requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.4.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
apscheduler==3.10.4
google-generativeai==0.5.4
openai==1.30.0
anthropic==0.28.0
email-validator==2.1.0
pymongo==4.8.0
httpx==0.27.2
bcrypt==4.0.1
